requirements.txt
# Core pipecat with needed service extras
pipecat-ai[nvidia,silero,elevenlabs,openai]==0.0.104

# Web server
fastapi==0.135.1
uvicorn==0.41.0
aiohttp==3.13.3
python-dotenv==1.2.2
loguru==0.7.3

# Telephony
plivo==4.59.6

# gRPC (NVIDIA NIM)
grpcio==1.67.1
grpcio-tools==1.67.1
nvidia-riva-client==2.25.0
